help完善

+++ app/controller/FoodService/food.py
@@ -3,6 +3,7 @@ from flask import request
 from app.db import User, Food, db
 
 class foodservice():
+
     def food(self):
         page = int(request.args.get('page'))
         #SecretKey = '0a6b58441e5069288e0f95939a2c4375'
@@ -37,6 +38,7 @@ class foodservice():
                 'state': '0',
                 'num' : 0
             }
+
     def creat(self):
         form = request.form
         UserPhone = form.get('UserPhone')
@@ -49,10 +51,10 @@ class foodservice():
                 UserId = u.UserId
                 p = Food.query.filter_by(UserId = UserId, State = 1).first()
                 if p is None:
-                    FoodArea = form.get('FoodArea')
-                    FoodInformation = form.args.get('FoodInformation')
+                    FoodArea = form.get('FoodArea').encode('utf-8')
+                    FoodInformation = form.args.get('FoodInformation').encode('utf-8')
                     FoodTime = form.args.get('FoodTime')
-                    FoodWay = form.args.get('FoodWay')
+                    FoodWay = form.args.get('FoodWay').encode('utf-8')
                     #FoodArea = 'hhh'
                     #FoodInformation = 'hhh'
                     #FoodTime = '2016-07-28 15:31:41'
@@ -73,15 +75,16 @@ class foodservice():
                     msg = '已有约会，创建失败'
             else:
                 state = '0'
-                msg = '创建失败'
+                msg = '请登录'
         else:
             state = '0'
-            msg = '创建失败'
+            msg = '请登录'
         array = {
             'state' : state,
             'msg' : msg
         }
         return array
+
     def get(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -112,6 +115,7 @@ class foodservice():
             'state' : state
         }
         return array
+
     def cancle(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -120,10 +124,14 @@ class foodservice():
             u = User.query.filter_by(UserPhone = UserPhone).first()
             if u.SecretKey == SecretKey:
                 f = Food.query.filter_by(FoodId = FoodId).first()
-                db.session.delete(f)
-                db.session.commit()
-                msg = '撤销成功'
-                state = '1'
+                if f.UserId == u.UserId:
+                    db.session.delete(f)
+                    db.session.commit()
+                    msg = '撤销成功'
+                    state = '1'
+                else:
+                    msg = '操作非法'
+                    state = '0'
             else:
                 msg = '请登录'
                 state = '0'
@@ -135,6 +143,7 @@ class foodservice():
             'msg' : msg
         }
         return array
+
     def myfood(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -173,6 +182,7 @@ class foodservice():
                 'state': '0'
             }
             return array
+
     def thisfood(self):
         FoodId = int(request.args.get('FoodId'))
         p = Food.query.filter_by(FoodId = FoodId).first()
@@ -192,9 +202,10 @@ class foodservice():
         else:
             array = {
                 'msg': '失败',
-                'state': '0',
+                'state': '0'
             }
         return array
+
     def accept(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -203,9 +214,13 @@ class foodservice():
             u = User.query.filter_by(UserPhone=UserPhone).first()
             if u.SecretKey == SecretKey:
                 f = Food.query.filter_by(FoodId=FoodId).first()
-                f.State = 1
-                msg = '成功'
-                state = '1'
+                if f.UserId == u.UserId:
+                    f.State = 0
+                    msg = '成功'
+                    state = '1'
+                else:
+                    msg = '操作非法'
+                    state = '0'
             else:
                 msg = '请登录'
                 state = '0'
@@ -217,6 +232,7 @@ class foodservice():
             'msg': msg
         }
         return array
+
     def refuse(self):
         UserPhone = request.args.get('UserPhone')
         SecretKey = request.args.get('SecretKey')
@@ -225,9 +241,13 @@ class foodservice():
             u = User.query.filter_by(UserPhone=UserPhone).first()
             if u.SecretKey == SecretKey:
                 f = Food.query.filter_by(FoodId=FoodId).first()
-                f.GetUser = None
-                msg = '已拒绝'
-                state = '1'
+                if f.UserId == u.UserId:
+                    f.GetUser = None
+                    msg = '已拒绝'
+                    state = '1'
+                else:
+                    msg = '操作非法'
+                    state = '0'
             else:
                 msg = '请登录'
                 state = '0'
